Peek SFN.Slot when handling P7 request messages
Add a function to get the SFN Slot of a P7 message without unpacking the whole message.

Rework is_nr_p7_request_in_window to not take the absolute of an unsigned value.
Instead, check if the result would be negative, and invert the operation if so.

Change pnf_handle_{msg}_request to peek SFN and Slot to determine whether the message is in window before unpacking the whole message