Check if slot type in P7 message is correct
Adds 'check_nr_nfapi_p7_slot_type()' which takes the SNF and Slot from a
(n)FAPI P7 message and checks whether the slot type is correct.  The
function prints a warning for the user that it received a message trying
to configure the wrong slot type upon failing the check.

Add this check on the handlers for DL_TTI, UL_TTI, UL_DCI and TX_DATA
requests