TODO can handle better? nr_generate_pdsch(): warn if no PDSCH message
nr_generate_pdsch() needs a PDU to encode a PDSCH message. At least in
nFAPI, it can happen that a DL_TTI.request message has been received
(configuring the PDSCH transmission), without the TX_data.request having
reached the PNF. This leads to a segfault without handling. As of now,
the mechanisms to prevent a triggering of PDSCH generation is not fully
implemented, so exit with an error message instead.

The solution would be to only actually trigger PDSCH if the
TX_data.request has reached the PNF.